Added separate implementations of certain parts of regolith taht are Windows specific.
