+++ regolith/compatibility_other_os.go
@@ -1,0 +1,10 @@
+//go:build !windows
+// +build !windows
+
+package regolith
+
+// copyFileSecurityInfo placeholder for a function which is necessary only
+// on Windows.
+func copyFileSecurityInfo(source string, target string) error {
+	return nil
+}
