+++ regolith/export.go
@@ -8,7 +8,6 @@ import (
 	"path/filepath"
 
 	"github.com/otiai10/copy"
-	"golang.org/x/sys/windows"
 )
 
 // GetExportPaths returns file paths for exporting behavior pack and
@@ -123,34 +122,6 @@ func ExportProject(profile Profile, name string) error {
 	}
 	err = editedFiles.Dump()
 	return err
-}
-
-// copyFileSecurityInfo copies the DACL info from source path to DACL of
-// the target path
-func copyFileSecurityInfo(source string, target string) error {
-	securityInfo, err := windows.GetNamedSecurityInfo(
-		source,
-		windows.SE_FILE_OBJECT,
-		windows.DACL_SECURITY_INFORMATION)
-	if err != nil {
-		return wrapError(
-			fmt.Sprintf("Unable to get security info of %q.", source), err)
-	}
-	dacl, _, err := securityInfo.DACL()
-	if err != nil {
-		return wrapError(
-			fmt.Sprintf("Unable to get DACL of %q.", source), err)
-	}
-	err = windows.SetNamedSecurityInfo(
-		target,
-		windows.SE_FILE_OBJECT,
-		windows.DACL_SECURITY_INFORMATION, nil, nil, dacl, nil,
-	)
-	if err != nil {
-		return wrapError(
-			fmt.Sprintf("Unable to set DACL of %q.", target), err)
-	}
-	return nil
 }
 
 // MoveOrCopy tries to move the the source to destination first and in case
