+++ test/export_windows_test.go
@@ -1,3 +1,6 @@
+//go:build windows
+// +build windows
+
 package test
 
 import (
